Align configuration contracts with Mermaid schema (#134)

+++ src/mermaiden/diagrams/requirement/configuration.py
@@ -10,8 +10,8 @@ class RequirementDiagramConfiguration(MermaidDiagramConfiguration):
     text_color: str = Field(default="#333", pattern=Text.pattern, description=Text.description)
     rect_border_size: str = Field(default="0.5px", pattern=Text.pattern, description=Text.description)
     rect_border_color: str = Field(default="#bbb", pattern=Text.pattern, description=Text.description)
-    rect_min_width: int = 200
-    rect_min_height: int = 200
-    font_size: int = 14
-    rect_padding: int = 10
-    line_height: int = 20
+    rect_min_width: float = 200
+    rect_min_height: float = 200
+    font_size: float = 14
+    rect_padding: float = 10
+    line_height: float = 20
